+++ packages/mangrove.js/CHANGELOG.md
@@ -5,6 +5,7 @@
 - `Market.{subscribe|consoleAsks|consoleBids|prettyPrint}` are no longer `async`
 - Removed `fromId` and `blockNumber` from `Market.BookOptions`
 - `Market.Offer.{prev|next}` are now `undefined` if there is no previous/next offer
+- `Market.getPivot` now throws `Error` if the order book cache is insufficient to determine a pivot
 
 # 0.0.8